requirements.txt
# Django Framework
Django==5.2.3

# PostgreSQL Database Adapter
psycopg2-binary==2.9.9

# WSGI HTTP Server for Production
gunicorn==21.2.0

# Static Files Serving for Production
whitenoise==6.6.0

# Environment Variables Management
python-dotenv==1.0.0

# Optional but Recommended for Production

# Redis Cache Backend (uncomment if using Redis)
# django-redis==5.4.0
# redis==5.0.1

# Security & Performance
django-ratelimit==4.1.0

# Monitoring & Error Tracking (uncomment if using Sentry)
# sentry-sdk==1.40.0

# Image Processing (if needed for media files)
# Pillow==10.1.0
